Refactor linked artifacts.
