+++ lib/hbc/artifact/hardlinked.rb
@@ -1,4 +1,6 @@
-class Hbc::Artifact::Hardlinked < Hbc::Artifact::Symlinked
+require "hbc/artifact/linked"
+
+class Hbc::Artifact::Hardlinked < Hbc::Artifact::Linked
   def self.islink?(path)
     return false unless path.respond_to?(:stat)
     path.stat.nlink > 1
@@ -11,7 +13,7 @@ class Hbc::Artifact::Hardlinked < Hbc::Artifact::Symlinked
   def create_filesystem_link(source, target)
     Pathname.new(target).dirname.mkpath
     @command.run!("/bin/ln", args: ["-hf", "--", source, target])
-    add_altname_metadata source, target
+    add_altname_metadata source, target.basename.to_s
   end
 
   def summarize_one_link(artifact_spec)
